refactor: refactors file architecture, fixes imports, changes icons
